refactoring: move generators to separate files
